+++ runtime/actors/misc.go
@@ -14,7 +14,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/hypermodeinc/modus/runtime/utils"
+	"github.com/hypermodeinc/modus/runtime/sentryutils"
 	goakt "github.com/tochemey/goakt/v3/actor"
 
 	"google.golang.org/protobuf/proto"
@@ -23,7 +23,7 @@ import (
 // Sends a message to an actor identified by its name.
 // Uses either Tell or RemoteTell based on whether the actor is local or remote.
 func tell(ctx context.Context, actorName string, message proto.Message) error {
-	span, ctx := utils.NewSentrySpanForCurrentFunc(ctx)
+	span, ctx := sentryutils.NewSpanForCurrentFunc(ctx)
 	defer span.Finish()
 
 	addr, pid, err := _actorSystem.ActorOf(ctx, actorName)
@@ -40,7 +40,7 @@ func tell(ctx context.Context, actorName string, message proto.Message) error {
 // Sends a message to an actor identified by its name, then waits for a response within the timeout duration.
 // Uses either Ask or RemoteAsk based on whether the actor is local or remote.
 func ask(ctx context.Context, actorName string, message proto.Message, timeout time.Duration) (response proto.Message, err error) {
-	span, ctx := utils.NewSentrySpanForCurrentFunc(ctx)
+	span, ctx := sentryutils.NewSpanForCurrentFunc(ctx)
 	defer span.Finish()
 
 	addr, pid, err := _actorSystem.ActorOf(ctx, actorName)