+++ runtime/models/bedrock.go
@@ -24,7 +24,7 @@ package models
 
 // func invokeAwsBedrockModel(ctx context.Context, model *manifest.ModelInfo, input string) (output string, err error) {
 
-// 	span, ctx := utils.NewSentrySpanForCurrentFunc(ctx)
+// 	span, ctx := sentryutils.NewSentrySpanForCurrentFunc(ctx)
 // 	defer span.Finish()
 
 // 	// NOTE: Bedrock support is experimental, and not advertised to users.