+++ CHANGELOG.md
@@ -7,6 +7,7 @@ NOTE: all releases may include dependency updates, not specifically mentioned
 ## UNRELEASED
 
 - feat: integrate try-as library [#912](https://github.com/hypermodeinc/modus/pull/912)
+- feat: improve sentry usage [#931](https://github.com/hypermodeinc/modus/pull/931)
 
 ## 2025-07-03 - Runtime v0.18.3